+++ src/test/java/org/folio/rest/api/StorageTestSuite.java
@@ -80,6 +80,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
   HoldingsSourceTest.class,
   InstanceDomainEventTest.class,
   InventoryViewTest.class,
+  InstanceSetTest.class,
   BoundWithStorageTest.class,
   ReindexJobRunnerTest.class,
   EffectiveLocationMigrationTest.class,